+++ Assets/Prefabs/Earth/Earth.prefab
@@ -427,7 +427,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 6703427069099394731}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: -0.03, z: 0}
+  m_LocalPosition: {x: 0, y: -0.01, z: 0}
   m_LocalScale: {x: 30, y: 2.3392043, z: 30}
   m_ConstrainProportionsScale: 0
   m_Children: []
